ValMapUpdateHandler: single-threaded `UpdateHandler` over ValMap
Single-threaded `UpdateHandler` implementation that stores updates
in a `ValMap`, locking the map for the entire duration of a commit.
After the commit is done, the map can be accessed from a different
thread.

Also re-arranges the Rust template so we can run `stack build` without
running DDlog.
