+++ rust/template/update_handler.rs
@@ -17,11 +17,35 @@ use super::*;
 
 use std::thread::*;
 use std::sync::mpsc::*;
-use std::sync::{Arc, Barrier, Mutex};
+use std::sync::{Arc, Barrier, Mutex, MutexGuard};
+use std::cell::RefCell;
+use super::valmap::*;
 
-pub trait UpdateHandler<V>: Send {
+/* Single-threaded (non-thread-safe callback)
+ */
+pub trait ST_CBFn<V>: Fn(RelId, &V, bool) {
+    fn clone_boxed(&self) -> Box<dyn ST_CBFn<V>>;
+}
+
+impl<T, V> ST_CBFn<V> for T
+where
+    V: Val,
+    T: 'static + Clone + Fn(RelId, &V, bool)
+{
+    fn clone_boxed(&self) -> Box<dyn ST_CBFn<V>> {
+        Box::new(self.clone())
+    }
+}
+
+impl <V: Val> Clone for Box<dyn ST_CBFn<V>> {
+    fn clone(&self) -> Self {
+        self.as_ref().clone_boxed()
+    }
+}
+
+pub trait UpdateHandler<V> {
     /* Returns a handler to be invoked on each output relation update. */
-    fn update(&self) -> Box<dyn CBFn<V>>;
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>>;
 
     /* Notifies the handler that a transaction_commit method is about to be
      * called. The handler has an opportunity to prepare to handle
@@ -36,30 +60,28 @@ pub trait UpdateHandler<V>: Send {
 /* Multi-threaded update handler that can be invoked from multiple DDlog
  * worker threads.
  */
-pub trait MTUpdateHandler<V>: UpdateHandler<V> + Sync {}
-impl <V: Val, H: UpdateHandler<V> + Sync + Send> MTUpdateHandler<V> for H {}
-
-type ExternCCallback = extern "C" fn(arg: libc::uintptr_t,
-                                     table: libc::size_t,
-                                     rec: *const record::Record,
-                                     polarity: bool);
+pub trait MTUpdateHandler<V>: UpdateHandler<V> + Sync + Send {
+    /* Returns a thread-safe handler to be invoked on each output
+     * relation update. */
+    fn mt_update_cb(&self) -> Box<dyn CBFn<V>>;
+}
 
 /* Rust magic to make `MTUpdateHandler` clonable.
  */
-pub trait DynUpdateHandler<V>: MTUpdateHandler<V> {
-   fn clone_boxed(&self) -> Box<dyn DynUpdateHandler<V>>;
+pub trait IMTUpdateHandler<V>: MTUpdateHandler<V> {
+   fn clone_boxed(&self) -> Box<dyn IMTUpdateHandler<V>>;
 }
 
-impl<T, V> DynUpdateHandler<V> for T
+impl<T, V> IMTUpdateHandler<V> for T
 where
    T: MTUpdateHandler<V> + Clone + 'static
 {
-   fn clone_boxed(&self) -> Box<dyn DynUpdateHandler<V>> {
+   fn clone_boxed(&self) -> Box<dyn IMTUpdateHandler<V>> {
        Box::new(self.clone())
    }
 }
 
-impl<V> Clone for Box<dyn DynUpdateHandler<V>> {
+impl<V> Clone for Box<dyn IMTUpdateHandler<V>> {
    fn clone(&self) -> Self {
        self.as_ref().clone_boxed()
    }
@@ -79,11 +101,18 @@ impl NullUpdateHandler
 
 impl <V: Val> UpdateHandler<V> for NullUpdateHandler
 {
-    fn update(&self) -> Box<dyn CBFn<V>> {
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>> {
         Box::new(|_,_,_|{})
     }
     fn before_commit(&self) {}
     fn after_commit(&self, _success: bool) {}
+}
+
+impl <V: Val> MTUpdateHandler<V> for NullUpdateHandler
+{
+    fn mt_update_cb(&self) -> Box<dyn CBFn<V>> {
+        Box::new(|_,_,_|{})
+    }
 }
 
 /* `UpdateHandler` implementation that invokes user-provided callback.
@@ -94,6 +123,11 @@ pub struct ExternCUpdateHandler {
     cb_arg:  libc::uintptr_t
 }
 
+type ExternCCallback = extern "C" fn(arg: libc::uintptr_t,
+                                     table: libc::size_t,
+                                     rec: *const record::Record,
+                                     polarity: bool);
+
 impl ExternCUpdateHandler
 {
     pub fn new(cb: ExternCCallback, cb_arg: libc::uintptr_t) -> Self {
@@ -103,13 +137,22 @@ impl ExternCUpdateHandler
 
 impl <V: Val + IntoRecord> UpdateHandler<V> for ExternCUpdateHandler
 {
-    fn update(&self) -> Box<dyn CBFn<V>> {
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>> {
         let cb = self.cb.clone();
         let cb_arg = self.cb_arg.clone();
         Box::new(move |relid, v, w|cb(cb_arg, relid, &v.clone().into_record() as *const record::Record, w))
     }
     fn before_commit(&self) {}
-    fn after_commit(&self, success: bool) {}
+    fn after_commit(&self, _success: bool) {}
+}
+
+impl <V: Val + IntoRecord> MTUpdateHandler<V> for ExternCUpdateHandler
+{
+    fn mt_update_cb(&self) -> Box<dyn CBFn<V>> {
+        let cb = self.cb.clone();
+        let cb_arg = self.cb_arg.clone();
+        Box::new(move |relid, v, w|cb(cb_arg, relid, &v.clone().into_record() as *const record::Record, w))
+    }
 }
 
 /* Multi-threaded `UpdateHandler` implementation that stores updates
@@ -117,44 +160,114 @@ impl <V: Val + IntoRecord> UpdateHandler<V> for ExternCUpdateHandler
  */
 #[derive(Clone)]
 pub struct MTValMapUpdateHandler {
-    db: sync::Arc<sync::Mutex<valmap::ValMap>>
+    db: Arc<Mutex<valmap::ValMap>>
 }
 
 impl MTValMapUpdateHandler
 {
-    pub fn new(db: sync::Arc<sync::Mutex<valmap::ValMap>>) -> Self {
+    pub fn new(db: Arc<Mutex<valmap::ValMap>>) -> Self {
         Self{db}
     }
 }
 
 impl UpdateHandler<Value> for MTValMapUpdateHandler
 {
-    fn update(&self) -> Box<dyn CBFn<Value>> {
+    fn update_cb(&self) -> Box<dyn ST_CBFn<Value>> {
         let db = self.db.clone();
         Box::new(move |relid, v, w|db.lock().unwrap().update(relid, v, w))
     }
     fn before_commit(&self) {}
-    fn after_commit(&self, success: bool) {}
+    fn after_commit(&self, _success: bool) {}
 }
 
-/* `UpdateHandler` implementation that chains multiple multi-threaded
- * handlers.
+impl MTUpdateHandler<Value> for MTValMapUpdateHandler
+{
+    fn mt_update_cb(&self) -> Box<dyn CBFn<Value>> {
+        let db = self.db.clone();
+        Box::new(move |relid, v, w|db.lock().unwrap().update(relid, v, w))
+    }
+}
+
+/* Single-threaded `UpdateHandler` implementation that stores updates
+ * in a `ValMap`, locking the map for the entire duration of a commit.
+ * After the commit is done, the map can be accessed from a different
+ * thread.
  */
 #[derive(Clone)]
-pub struct ChainedDynUpdateHandler<V: Val> {
-    handlers: Vec<Box<dyn DynUpdateHandler<V>>>
+pub struct ValMapUpdateHandler {
+    db: Arc<Mutex<ValMap>>,
+    /* Stores pointer to `MutexGuard` between `before_commit()` and
+    * `after_commit()`.  This has to be unsafe, because Rust does
+    * not let us express a borrow from a field of the same struct. */
+    locked: Arc<RefCell<*mut libc::c_void>>
 }
 
-impl <V: Val> ChainedDynUpdateHandler<V> {
-    pub fn new(handlers: Vec<Box<dyn DynUpdateHandler<V>>>) -> Self {
+impl Drop for ValMapUpdateHandler {
+    /* Release the mutex if still held. */
+    fn drop<'a>(&'a mut self) {
+        let guard_ptr = self.locked.replace(ptr::null_mut()) as *mut MutexGuard<'a, ValMap>;
+        if guard_ptr != ptr::null_mut() {
+            let _guard: Box<MutexGuard<'_, ValMap>> = unsafe { Box::from_raw(guard_ptr) };
+        }
+    }
+}
+
+impl ValMapUpdateHandler
+{
+    pub fn new(db: Arc<Mutex<ValMap>>) -> Self {
+        Self{db, locked: Arc::new(RefCell::new(ptr::null_mut()))}
+    }
+}
+
+impl UpdateHandler<Value> for ValMapUpdateHandler
+{
+    fn update_cb(&self) -> Box<dyn ST_CBFn<Value>> {
+        let handler = self.clone();
+        Box::new(move |relid, v, w|{
+            let guard_ptr = handler.locked.borrow();
+            /* `update_cb` can only be called between `before_commit()` and
+             * `after_commit()` */
+            assert_ne!(*guard_ptr, ptr::null_mut());
+            let mut guard: Box<MutexGuard<'_, ValMap>> = unsafe {
+                Box::from_raw(*guard_ptr as *mut MutexGuard<'_, ValMap>)
+            };
+            guard.update(relid, v, w);
+            Box::into_raw(guard);
+        })
+    }
+    fn before_commit(&self) {
+        let guard = Box::into_raw(Box::new(self.db.lock().unwrap())) as *mut libc::c_void;
+        let old = self.locked.replace(guard);
+        assert_eq!(old, ptr::null_mut());
+    }
+    fn after_commit(&self, _success: bool) {
+        let guard_ptr = self.locked.replace(ptr::null_mut());
+        assert_ne!(guard_ptr, ptr::null_mut());
+        let _guard = unsafe {
+            Box::from_raw(guard_ptr as *mut MutexGuard<'_, ValMap>)
+        };
+        /* Lock will be released when `_guard` goes out of scope. */
+    }
+}
+
+
+/* `UpdateHandler` implementation that chains multiple single-threaded
+ * handlers.
+ */
+pub struct ChainedUpdateHandler<V: Val> {
+    handlers: Vec<Box<dyn UpdateHandler<V>>>
+}
+
+impl <V: Val> ChainedUpdateHandler<V> {
+    pub fn new(handlers: Vec<Box<dyn UpdateHandler<V>>>) -> Self {
         Self{ handlers }
     }
 }
 
-impl <V: Val> UpdateHandler<V> for ChainedDynUpdateHandler<V>
+impl <V: Val> UpdateHandler<V> for ChainedUpdateHandler<V>
 {
-    fn update(&self) -> Box<dyn CBFn<V>> {
-        let cbs: Vec<Box<dyn CBFn<V>>> = self.handlers.iter().map(|h|h.update()).collect();
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>> {
+        let cbs: Vec<Box<dyn ST_CBFn<V>>> = self.handlers.iter().map(|h|h.update_cb()).collect();
         Box::new(move |relid, v, w| {
             for cb in cbs.iter() {
                 cb(relid, v, w);
@@ -174,6 +287,55 @@ impl <V: Val> UpdateHandler<V> for ChainedDynUpdateHandler<V>
     }
 }
 
+/* `UpdateHandler` implementation that chains multiple multi-threaded
+ * handlers.
+ */
+#[derive(Clone)]
+pub struct MTChainedUpdateHandler<V: Val> {
+    handlers: Vec<Box<dyn IMTUpdateHandler<V>>>
+}
+
+impl <V: Val> MTChainedUpdateHandler<V> {
+    pub fn new(handlers: Vec<Box<dyn IMTUpdateHandler<V>>>) -> Self {
+        Self{ handlers }
+    }
+}
+
+impl <V: Val> UpdateHandler<V> for MTChainedUpdateHandler<V>
+{
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>> {
+        let cbs: Vec<Box<dyn ST_CBFn<V>>> = self.handlers.iter().map(|h|h.update_cb()).collect();
+        Box::new(move |relid, v, w| {
+            for cb in cbs.iter() {
+                cb(relid, v, w);
+            }
+        })
+    }
+    fn before_commit(&self) {
+        for h in self.handlers.iter() {
+            h.before_commit();
+        }
+    }
+
+    fn after_commit(&self, success: bool) {
+        for h in self.handlers.iter() {
+            h.after_commit(success);
+        }
+    }
+}
+
+impl <V: Val> MTUpdateHandler<V> for MTChainedUpdateHandler<V>
+{
+    fn mt_update_cb(&self) -> Box<dyn CBFn<V>> {
+        let cbs: Vec<Box<dyn CBFn<V>>> = self.handlers.iter().map(|h|h.mt_update_cb()).collect();
+        Box::new(move |relid, v, w| {
+            for cb in cbs.iter() {
+                cb(relid, v, w);
+            }
+        })
+    }
+}
+
 /* `UpdateHandler` implementation that handles updates in a separate
  * worker thread.
  */
@@ -190,7 +352,7 @@ enum Msg<V: Val> {
 #[derive(Clone)]
 pub struct ThreadUpdateHandler<V: Val> {
     /* Channel to worker thread. */
-    msg_channel: Arc<Mutex<Sender<Msg<V>>>>,
+    msg_channel: Arc<Mutex<SyncSender<Msg<V>>>>,
 
     /* Barrier to synchronize completion of transaction with worker. */
     commit_barrier: Arc<Barrier>
@@ -198,12 +360,15 @@ pub struct ThreadUpdateHandler<V: Val> {
 
 impl <V: Val> ThreadUpdateHandler<V>
 {
-    pub fn new<H: UpdateHandler<V> + 'static>(handler: H) -> Self {
-        let (tx_msg_channel, rx_message_channel) = channel();
+    pub fn new<F>(handler_generator: F, queue_capacity: usize) -> Self 
+        where F: FnOnce() -> Box<dyn UpdateHandler<V>> + Send + 'static
+    {
+        let (tx_msg_channel, rx_message_channel) = sync_channel(queue_capacity);
         let commit_barrier = Arc::new(Barrier::new(2));
         let commit_barrier2 = commit_barrier.clone();
         spawn(move || {
-            let update_cb = handler.update();
+            let handler = handler_generator();
+            let update_cb = handler.update_cb();
             loop {
                 match rx_message_channel.recv() {
                     Ok(Msg::Update{relid, v, w}) => {
@@ -246,7 +411,7 @@ impl<V: Val> Drop for ThreadUpdateHandler<V> {
 
 impl <V: Val + IntoRecord> UpdateHandler<V> for ThreadUpdateHandler<V>
 {
-    fn update(&self) -> Box<dyn CBFn<V>> {
+    fn update_cb(&self) -> Box<dyn ST_CBFn<V>> {
         let channel = self.msg_channel.lock().unwrap().clone();
         Box::new(move |relid, v, w| {
             channel.send(Msg::Update{relid, v: v.clone(), w});
@@ -260,5 +425,15 @@ impl <V: Val + IntoRecord> UpdateHandler<V> for ThreadUpdateHandler<V>
 
         /* Wait for all queued updates to get processed by worker. */
         self.commit_barrier.wait();
+    }
+}
+
+impl <V: Val + IntoRecord> MTUpdateHandler<V> for ThreadUpdateHandler<V>
+{
+    fn mt_update_cb(&self) -> Box<dyn CBFn<V>> {
+        let channel = self.msg_channel.lock().unwrap().clone();
+        Box::new(move |relid, v, w| {
+            channel.send(Msg::Update{relid, v: v.clone(), w});
+        })
     }
 }
